# backend/docs/swagger.yaml
basePath: /
definitions:
  customerrors.Wrapper:
    properties:
      description:
        type: string
      error:
        description: 'TODO: Remove id from Wrapper'
      id:
        type: integer
    type: object
  dto.NewProfileDTO:
    properties:
      age:
        maximum: 100
        minimum: 18
        type: integer
      last_name:
        type: string
      name:
        type: string
    required:
    - age
    - last_name
    - name
    type: object
  dto.PatchProfileDTO:
    properties:
      age:
        maximum: 100
        minimum: 18
        type: integer
      last_name:
        type: string
      name:
        type: string
    type: object
  dto.ProfileDTO:
    properties:
      age:
        type: integer
      id:
        type: integer
      last_name:
        type: string
      name:
        type: string
    type: object
info:
  contact: {}
  description: My first API
  license:
    name: MIT
  title: CRUD app
  version: "1.0"
paths:
  /calculator/add/{num1}/{num2}:
    get:
      consumes:
      - application/json
      parameters:
      - description: First number
        in: path
        name: num1
        required: true
        type: integer
      - description: Second number
        in: path
        name: num2
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
      summary: Addition of two numbers
      tags:
      - Calculator
  /calculator/count:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
      summary: Count of operations
      tags:
      - Calculator
  /calculator/division/{num1}/{num2}:
    get:
      consumes:
      - application/json
      parameters:
      - description: First number
        in: path
        name: num1
        required: true
        type: integer
      - description: Second number
        in: path
        name: num2
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
      summary: Division of two numbers
      tags:
      - Calculator
  /calculator/sub/{num1}/{num2}:
    get:
      consumes:
      - application/json
      parameters:
      - description: First number
        in: path
        name: num1
        required: true
        type: integer
      - description: Second number
        in: path
        name: num2
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
      summary: Subtraction of two numbers
      tags:
      - Calculator
  /profile:
    post:
      consumes:
      - application/json
      parameters:
      - description: New profile
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.NewProfileDTO'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: int
        "400":
          description: Bad Request
      summary: Create a new profile
      tags:
      - Profile
  /profile/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Profile id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Profile not found
          schema:
            $ref: '#/definitions/customerrors.Wrapper'
      summary: Delete a profile
      tags:
      - Profile
    get:
      consumes:
      - application/json
      parameters:
      - description: Profile id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/dto.ProfileDTO'
        "400":
          description: Bad Request
        "404":
          description: Profile not found
          schema:
            $ref: '#/definitions/customerrors.Wrapper'
      summary: Get a profile
      tags:
      - Profile
    patch:
      consumes:
      - application/json
      parameters:
      - description: Profile id
        in: path
        name: id
        required: true
        type: integer
      - description: New profile
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.PatchProfileDTO'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/dto.ProfileDTO'
        "400":
          description: Bad Request
        "404":
          description: Profile not found
          schema:
            $ref: '#/definitions/customerrors.Wrapper'
      summary: Patch a profile
      tags:
      - Profile
    put:
      consumes:
      - application/json
      parameters:
      - description: Profile id
        in: path
        name: id
        required: true
        type: integer
      - description: New profile
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.NewProfileDTO'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/dto.ProfileDTO'
        "400":
          description: Bad Request
        "404":
          description: Profile not found
          schema:
            $ref: '#/definitions/customerrors.Wrapper'
      summary: Update a profile
      tags:
      - Profile
schemes:
- http
securityDefinitions:
  JWT:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"